sql-injection/requirements.txt
Flask==2.1.0
Flask-Cors==3.0.8
psycopg2-binary==2.8.5
SQLAlchemy==1.3.18
numpy==1.20.3
